Setup a new angular project
