+++ iot/angular/iot/README.md
@@ -1,4 +1,4 @@
-# Angular
+# Iot
 
 This project was generated with [Angular CLI](https://github.com/angular/angular-cli) version 6.0.0.
