+++ iot/angular/iot/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "angular",
+  "name": "iot",
   "version": "0.0.0",
   "scripts": {
     "ng": "ng",
@@ -11,29 +11,25 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^6.0.7",
-    "@angular/cdk": "6.3.1",
-    "@angular/common": "^6.0.7",
-    "@angular/compiler": "^6.0.7",
-    "@angular/core": "^6.0.7",
-    "@angular/forms": "^6.0.7",
-    "@angular/http": "^6.0.7",
-    "@angular/material": "^6.2.1",
-    "@angular/platform-browser": "^6.0.7",
-    "@angular/platform-browser-dynamic": "^6.0.7",
-    "@angular/router": "^6.0.7",
-    "chart.js": "^2.7.2",
+    "@angular/animations": "^6.0.0",
+    "@angular/common": "^6.0.0",
+    "@angular/compiler": "^6.0.0",
+    "@angular/core": "^6.0.0",
+    "@angular/forms": "^6.0.0",
+    "@angular/http": "^6.0.0",
+    "@angular/platform-browser": "^6.0.0",
+    "@angular/platform-browser-dynamic": "^6.0.0",
+    "@angular/router": "^6.0.0",
     "core-js": "^2.5.4",
-    "hammerjs": "^2.0.8",
-    "rxjs": "^6.2.1",
+    "rxjs": "^6.0.0",
     "zone.js": "^0.8.26"
   },
   "devDependencies": {
-    "@angular/compiler-cli": "^6.0.7",
+    "@angular/compiler-cli": "^6.0.0",
     "@angular-devkit/build-angular": "~0.6.0",
     "typescript": "~2.7.2",
-    "@angular/cli": "6.0.8",
-    "@angular/language-service": "^6.0.7",
+    "@angular/cli": "~6.0.0",
+    "@angular/language-service": "^6.0.0",
     "@types/jasmine": "~2.8.6",
     "@types/jasminewd2": "~2.0.3",
     "@types/node": "~8.9.4",
