+++ iot/angular/iot/src/app/app.component.spec.ts
