+++ iot/angular/iot/src/index.html
@@ -1,19 +1,14 @@
 <!doctype html>
 <html lang="en">
-
 <head>
-  <link href="https://fonts.googleapis.com/icon?family=Material+Icons" rel="stylesheet">
-  <link href="https://fonts.googleapis.com/css?family=Roboto:300,400,500" rel="stylesheet">
   <meta charset="utf-8">
-  <title>IOT</title>
+  <title>Iot</title>
   <base href="/">
 
   <meta name="viewport" content="width=device-width, initial-scale=1">
   <link rel="icon" type="image/x-icon" href="favicon.ico">
 </head>
-
 <body>
   <app-root></app-root>
 </body>
-
 </html>
