+++ iot/angular/iot/src/karma.conf.js
