+++ iot/angular/iot/src/main.ts
@@ -4,8 +4,6 @@ import { platformBrowserDynamic } from '@angular/platform-browser-dynamic';
 import { AppModule } from './app/app.module';
 import { environment } from './environments/environment';
 
-import 'hammerjs';
-
 if (environment.production) {
   enableProdMode();
 }
