--- iot/angular/src/app/app-routing.module.ts
@@ -1,22 +1,0 @@
-import { NgModule } from '@angular/core';
-import { RouterModule, Routes } from '@angular/router';
-
-import { HomeComponent } from './home/home.component';
-import { StatusComponent } from './status/status.component';
-import { HistoryComponent } from './history/history.component';
-import { NotfoundComponent } from './notfound/notfound.component';
-import { LoginComponent } from './login/login.component';
-
-const routes: Routes = [
-  { path: '', component: HomeComponent },
-  { path: 'status', component: StatusComponent },
-  { path: 'history', component: HistoryComponent },
-  { path: 'login', component: LoginComponent },
-  { path: '**', component: NotfoundComponent},
-];
-
-@NgModule({
-  imports: [ RouterModule.forRoot(routes) ],
-  exports: [ RouterModule ]
-})
-export class AppRoutingModule {}
