--- iot/angular/src/app/app.component.html
@@ -1,28 +1,0 @@
-<div class="container" [class.is-mobile]="mobileQuery.matches">
-  <mat-toolbar color="primary" class="toolbar">
-    <button mat-icon-button (click)="snav.toggle()"><mat-icon>menu</mat-icon></button>
-    <h1 class="app-name">IOT</h1>
-  </mat-toolbar>
-
-  <mat-sidenav-container class="sidenav-container"
-                         [style.marginTop.px]="mobileQuery.matches ? 56 : 0">
-    <mat-sidenav #snav [mode]="mobileQuery.matches ? 'over' : 'side'"
-                 [fixedInViewport]="mobileQuery.matches" fixedTopGap="56">
-      <mat-nav-list>
-          <mat-list-item>
-              <a class="nav-link" routerLink="/">Home</a>
-          </mat-list-item>
-          <mat-list-item>
-              <a class="nav-link" routerLink="status">Status</a>
-          </mat-list-item>
-          <mat-list-item>
-              <a class="nav-link" routerLink="history">History</a>
-          </mat-list-item>
-      </mat-nav-list>
-    </mat-sidenav>
-
-    <mat-sidenav-content>
-      <router-outlet></router-outlet>
-    </mat-sidenav-content>
-  </mat-sidenav-container>
-</div>
