--- iot/angular/src/app/history/history.component.html
@@ -1,3 +1,0 @@
-<div *ngIf="chart">
-  <canvas id="canvas">{{ chart }}</canvas>
-</div>
