--- iot/angular/src/app/home/home.component.html
@@ -1,3 +1,0 @@
-<p>
-  Welcome!
-</p>
