--- iot/angular/src/app/login/login.component.html
@@ -1,3 +1,0 @@
-<p>
-  login works!
-</p>
