--- iot/angular/src/app/notfound/notfound.component.html
@@ -1,8 +1,0 @@
-<p>你要查看的网页已经消失到另一个次元了。</p>
-<p>
-  请尝试以下操作：
-  <br /> 1. 如果您已经在地址栏中输入该网页的地址，请确认其拼写正确；
-  <br /> 2. 点击首页进入
-  <a routerLink="/">https://hemengyang.tk/</a> 主页；
-  <br /> 3. 单击"后退"按钮，尝试其他链接；
-</p>
