--- iot/angular/src/app/status.ts
@@ -1,7 +1,0 @@
-export class Status {
-    time: string;
-    temperature: number;
-    relativeHumidity: number;
-    relay1Status: boolean;
-    relay2Status: boolean;
-  }
