--- iot/angular/src/app/status/status.component.html
@@ -1,48 +1,0 @@
-<mat-list>
-  <mat-list-item>
-    <span class="h5">Time:&nbsp;</span>
-    <span class="data">{{status?.time | date: 'medium':'+0800'}}</span>
-  </mat-list-item>
-  <mat-divider></mat-divider>
-  <mat-list-item>
-    <span class="h5">Temperature:&nbsp;</span>
-    <span class="data">{{status?.temperature}}</span>
-  </mat-list-item>
-  <mat-divider></mat-divider>
-  <mat-list-item>
-    <span class="h5">Relative Humidity:&nbsp;</span>
-    <span class="data">{{status?.relativeHumidity}}</span>
-  </mat-list-item>
-  <mat-divider></mat-divider>
-  <mat-list-item>
-    <span class="h5">Relay1 Status:&nbsp;</span>
-    <span class="data">{{status?.relay1Status}}</span>
-  </mat-list-item>
-  <mat-divider></mat-divider>
-  <mat-list-item>
-    <span class="h5">Relay2 Status:&nbsp;</span>
-    <span class="data">{{status?.relay2Status}}</span>
-  </mat-list-item>
-</mat-list>
-<!-- <div class="p-2">
-    <span class="h5">Relay1 Status:&nbsp;</span>
-    <div class="btn-group" (click)="changeRelay1()">
-      <label class="btn btn-primary waves-light btn-sm" [(ngModel)]="relay1" mdbRadio="ON" mdbRippleRadius>
-          ON
-      </label>
-      <label class="btn btn-primary waves-light btn-sm" [(ngModel)]="relay1" mdbRadio="OFF" mdbRippleRadius>
-          OFF
-      </label>
-    </div>
-  </div>
-  <div class="p-2">
-    <span class="h5">Relay2 Status:&nbsp;</span>
-    <div class="btn-group" (click)="changeRelay2()">
-      <label class="btn btn-primary waves-light btn-sm" [(ngModel)]="relay2" mdbRadio="ON" mdbRippleRadius>
-          ON
-      </label>
-      <label class="btn btn-primary waves-light btn-sm" [(ngModel)]="relay2" mdbRadio="OFF" mdbRippleRadius>
-          OFF
-      </label>
-    </div>
-  </div> -->
